+++ mlflow/server/js/src/MlflowRouter.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useMemo, useState } from 'react';
 import { LegacySkeleton, useDesignSystemTheme } from '@databricks/design-system';
+import { useDocumentTitle } from '@databricks/web-shared/routing';
 
 import ErrorModal from './experiment-tracking/components/modals/ErrorModal';
 import AppErrorBoundary from './common/components/error-boundaries/AppErrorBoundary';
@@ -9,11 +10,10 @@ import {
   Outlet,
   createLazyRouteElement,
   useParams,
+  useGetRouteTitle,
 } from './common/utils/RoutingUtils';
 import { MlflowHeader } from './common/components/MlflowHeader';
 import { useDarkThemeContext } from './common/contexts/DarkThemeContext';
-import { useDocumentTitle } from './common/hooks/useDocumentTitle';
-import type { DocumentTitleHandle } from './common/hooks/useDocumentTitle';
 
 // Route definition imports:
 import { getRouteDefs as getExperimentTrackingRouteDefs } from './experiment-tracking/route-defs';
@@ -28,7 +28,9 @@ import { MlflowSidebar } from './common/components/MlflowSidebar';
  */
 const MlflowRootRoute = () => {
   useInitializeExperimentRunColors();
-  useDocumentTitle();
+
+  const routeTitle = useGetRouteTitle();
+  useDocumentTitle({ title: routeTitle });
 
   const [showSidebar, setShowSidebar] = useState(true);
   const { theme } = useDesignSystemTheme();
