+++ mlflow/server/js/src/common/route-defs.ts
@@ -1,6 +1,5 @@
 import PageNotFoundView from './components/PageNotFoundView';
-import { createMLflowRoutePath, createRouteElement } from './utils/RoutingUtils';
-import type { DocumentTitleHandle } from './hooks/useDocumentTitle';
+import { createMLflowRoutePath, createRouteElement, type DocumentTitleHandle } from './utils/RoutingUtils';
 
 /**
  * Common route definitions. For the time being it's 404 page only.
