+++ mlflow/server/js/src/shared/web-shared/routing/index.tsx
@@ -1,0 +1,1 @@
+export { useDocumentTitle } from './useDocumentTitle';
